refactor(docs): extract shared leaf helpers into lib/doc-utils.ts
First cut of the docs generator split (audit-2026-06 WS7). Moves the
dependency-free leaf helpers every renderer uses into a new lib/doc-utils.ts:
writeStableFile (frontmatter-aware writer), docsPath + docsBasePath, the
HTML/markdown escapers (escapeHtml, escapeMd, escapeTableCell,
escapeMarkdownTableCell), generatedPageNote, and flattenObject.

This is the foundation the subsequent renderer-cluster extractions import.
generate.ts imports them back (+ docsBasePath for its docsUrl) and shed its
now-unused mkdirSync/writeFileSync/dirname imports.

Pure relocation — bodies byte-for-byte unchanged. generate.ts 2,358 -> 2,308.
Verified: eslint (changed files) clean, and docs:generate output is
BYTE-IDENTICAL to the pre-refactor baseline (deterministic generator; the
relocation gate, since the script is dev-only tooling not in the binary).
